Move admin models into the main models package.
Since they're already in their own module a distinct package is not
needed.
