+++ modules/admin/app/models/BatchDeleteTask.scala
@@ -1,4 +1,4 @@
-package models.admin
+package models
 
 import play.api.data.Form
 import play.api.data.Forms._
