+++ modules/admin/app/models/OaiPmhConfig.scala
@@ -1,4 +1,4 @@
-package models.admin
+package models
 
 import play.api.data.Form
 import play.api.data.Forms._
