+++ modules/admin/app/views/admin/repository/harvest.scala.html
@@ -1,8 +1,8 @@
-@(item: Repository, f: play.api.data.Form[models.admin.OaiPmhConfig], action: Call)(implicit userOpt: Option[UserProfile], req: RequestHeader, globalConfig: global.GlobalConfig, messages: Messages, md: views.MarkdownRenderer, prefs: utils.SessionPrefs, flash: Flash)
+@(item: Repository, f: play.api.data.Form[OaiPmhConfig], action: Call)(implicit userOpt: Option[UserProfile], req: RequestHeader, globalConfig: global.GlobalConfig, messages: Messages, md: views.MarkdownRenderer, prefs: utils.SessionPrefs, flash: Flash)
 
 @implicitField = @{ views.html.helper.FieldConstructor(formHelpers.fieldTemplate.f) }
 
-@import models.admin.OaiPmhConfig._
+@import OaiPmhConfig._
 
 @views.html.admin.layout.rightSidebar(Messages("oaipmh"), breadcrumbs = views.html.admin.common.breadcrumbs(List(item))) {
     @helper.form(action = action, 'enctype -> "multipart/form-data") {
